+++ fresh.gen.ts
@@ -20,6 +20,7 @@ import * as $markdown_index from "./routes/markdown/index.tsx";
 import * as $panel_panel from "./routes/panel/panel.tsx";
 import * as $text_text from "./routes/text/text.tsx";
 import * as $DemoButton from "./islands/DemoButton.tsx";
+import * as $DemoInput from "./islands/DemoInput.tsx";
 import * as $KvAutocomplete from "./islands/KvAutocomplete.tsx";
 import { type Manifest } from "$fresh/server.ts";
 
@@ -45,6 +46,7 @@ const manifest = {
   },
   islands: {
     "./islands/DemoButton.tsx": $DemoButton,
+    "./islands/DemoInput.tsx": $DemoInput,
     "./islands/KvAutocomplete.tsx": $KvAutocomplete,
   },
   baseUrl: import.meta.url,